feat: move verification step into pickup (#128)
This removes the verification queue in favour of doing CAR verification
as part of the pickup worker. We're seeing the verification step consume
very little mem / disk / cpu, and the improvement here is we can get rid
of a whole extra container service and queue, and dead-letter queue etc.

We now wire up pickup with an S3Uploader that takes a `validationBucket`
and a `destinationBucket`. It writes the upload to the
`validationBucket` and then retrieves it again, streaming the body
through `CarBlockIterator` and the `LinkIndexer` and if the DAG is
complete the CAR is copied to the destinationBucket. This is the same
logic as used in the old external validation task, but now as the final
step of the pickup s3 upload. The validation bucket is configured to
auto expire uploads after a few hours, so aws can delete them at it's
leisure.

Some effort has been made to make the PickupStack easier to understand
again by reducing the number of places you have to look to figure out
what config will be applied.

---------

Signed-off-by: Oli Evans <oli@protocol.ai>