feat: calculate car cid and size during validation (#137)
Update car key to `${carCid}/${carCid}.car`, so we can start writing
them to carpark.

Calculate the sha256 of the car bytes from the stream as we produce the
linkdex report to determine if the car is complete / valid

Log total bytes per CAR as `carSize`. Use this to calculate the
throughput in loki. Needed as we wont be able to use the size of the
bucket once we start writing to the shared carpark.

Fixes https://github.com/web3-storage/pickup/issues/136

TODO
- [ ] update the production and staging env to write to carpark bucket.

License: MIT

---------

Signed-off-by: Oli Evans <oli@protocol.ai>