feat: write to carpark (#139)
Write to the existing carpark bucket rather than creating a dedicated
one for pickup.

Updater the pin status on successful write to s3 from the pickup worker.
Previously we'd listen for object_created events and trigger a lambda to
update the pin status, but we're moving to a shared bucket, so handle it
in the worker.

Key changes
- Update sst stack to adopt bucket by name rather than create it's own
- Delete the pinUpdate queue, as we now update the pin directly from the
pickup worker
- Improve the error handling so we fail pins that are too large rather
than retying them... as we are updating the pin status from the worker
now this is feasible. Previously we'd only mark a pin as failed using
the dead-letter queue handler.
- Fix pickup tests to verify pin status update in the dynamo table.
- Pass the abort signal for the car fetch to the s3 upload so we abort
the upload correctly. this fixes timeout and error handling issues we
were seeing in the tests where the fetch timeout and chunk timeout would
trigger the abort signal, but an error was not thrown immediately, which
was making tests hang.


Part of #136

---------

Signed-off-by: Oli Evans <oli@protocol.ai>
Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>